% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{estimate_area}
\alias{estimate_area}
\alias{estimate_area,plot}
\title{Estimation of a shape area function}
\usage{
estimate_area(B = 5000, seed = 10)
}
\arguments{
\item{B}{Number of random data points (x,y) to generate.}

\item{seed}{Pseudorandom number generator.}
}
\value{
A \code{list} with estimation of the area of S and the simulated data points.
}
\description{
Estimates the area of a shape S by simulating random coordinates in a one by one square.
}
\examples{
estimate_area(B = 3456, seed = 100)
}
