% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{plot.area}
\alias{plot.area}
\title{Plotting of a shape area function}
\usage{
\method{plot}{area}(x, ...)
}
\arguments{
\item{x}{The return value from the estimation of the area of the shape function.}
}
\value{
Plot of the shape S and the simulated data points (within the shape and outside the shape).
}
\description{
Plots the area of a shape S by simulating random coordinates in a one by one square.
}
\examples{
plot.area(estimate_area(B = 1000, seed = 10))
}
\author{
Virany Kho, Colin Steffe, Guillaume Bilocq, Daniel Welz and Jasmine Mawjee
}
