% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{plot_area}
\alias{plot_area}
\title{Plotting of a shape area function}
\usage{
plot_area(x)
}
\arguments{
\item{x}{Return value from the estimation of the area of the shape function.}
}
\value{
Shape S and the simulated data points (within the shape and outside the shape).
}
\description{
Plot the area of a shape S by simulating random coordinates in a one by one square.
}
\examples{
plot(x = estimate_area())
}
