\name{Estimation and plot of a shape area}
\title{Estimate and plot the area of a shape}
\description{Estimates and plots the area of a shape S by simulating random coordinates in a one by one square}
\usage{
estimate_area(B = 5000, seed = 10)
plot.area(x)}
\arguments{
\item{B}{A number of random data points (x,y) to generate.}
\item{seed}{Initializes a pseudorandom number generator.}
\item{x}{Returns value from the estimation of the area of the shape.}
}


