% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{estimate_area}
\alias{estimate_area}
\alias{estimate_area,plot.area}
\title{Estimation of a shape area function}
\usage{
estimate_area(B = 5000, seed = 10)
}
\arguments{
\item{B}{A number of random data points (x,y) to generate.}

\item{seed}{Initializes a pseudorandom number generator.}
}
\value{
Estimation of the area of the shape S.
}
\description{
Estimates the area of a shape S by simulating random coordinates in a one by one square.
}
\examples{
estimate_area(B = 5000, seed = 10)
}
\author{
Virany Kho, Colin Steffe, Guillaume Bilocq, Daniel Welz and Jasmine Mawjee
}
